--- src/api_conformance/mod.rs
@@ -1,16 +1,0 @@
-#[cfg(test)]
-mod btreemap;
-#[cfg(test)]
-mod btreeset;
-#[cfg(test)]
-mod min_heap;
-#[cfg(test)]
-mod vec;
-
-use std::cell::RefCell;
-use std::rc::Rc;
-
-#[cfg(test)]
-pub(crate) fn make_memory() -> Rc<RefCell<std::vec::Vec<u8>>> {
-    Rc::new(RefCell::new(std::vec::Vec::new()))
-}
